Improved physics (slightly).
I'm trying to reduce the car bounciness. By changing the physics engine to "Godot Physics" it seems some of it was reduced a lot. Also, lerping the car's y component helps. I think there is room to improve much more by changing the bounciness of the physics materials for the track and floor.